+++ bike-journey-viewer/.eslintrc.cjs
@@ -20,5 +20,6 @@ module.exports = {
       { allowConstantExport: true },
     ],
     'import/no-extraneous-dependencies': 0,
+    'react/react-in-jsx-scope': 0,
   },
 };