+++ bike-journey-viewer/src/App.tsx
@@ -1,29 +1,34 @@
-import React, { useState } from 'react';
-import reactLogo from './assets/react.svg';
+import React from 'react';
+import { Route, Routes, useNavigate } from 'react-router-dom';
+import Toolbar from './features/toolbar/Toolbar';
 import './App.css';
 
 function App() {
-  const [count, setCount] = useState(0);
+  const navigate = useNavigate();
+
+  const handleMainClick = () => {
+    navigate('/');
+  };
+
+  const handleStationClick = () => {
+    navigate('/stations');
+  };
+
+  const handleGitHubClick = () => {
+    window.location.href = 'https://github.com/wanc9/bike-journey-viewer';
+  };
 
   return (
     <>
-      <div>
-        <a href="https://react.dev" target="_blank" rel="noreferrer">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button type="button" onClick={() => setCount(count + 1)}>
-          count is {count}
-        </button>
-        <p>
-          Edit <code>src/App.tsx</code> and save to test HMR
-        </p>
-      </div>
-      <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p>
+      <Toolbar
+        handleMainClick={handleMainClick}
+        handleStationClick={handleStationClick}
+        handleGitHubClick={handleGitHubClick}
+      />
+      <Routes>
+        <Route path="/" element={<h1>Welcome!</h1>} />
+        <Route path="/stations" element={<h1>Stations</h1>} />
+      </Routes>
     </>
   );
 }